extract GetTypesScopedByTestClass implementation again
