+++ src/StorageQueues.Tests/DefaultEndpoint.cs
@@ -1,15 +1,11 @@
 ﻿namespace StorageQueues.Tests
 {
     using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Reflection;
     using System.Threading.Tasks;
     using NServiceBus;
     using NServiceBus.AcceptanceTesting.Customization;
     using NServiceBus.AcceptanceTesting.Support;
     using NServiceBus.Features;
-    using NServiceBus.Hosting.Helpers;
 
     class DefaultEndpoint : IEndpointSetupTemplate
     {
@@ -20,7 +16,7 @@
         {
             var configuration = new EndpointConfiguration(endpointConfiguration.EndpointName);
 
-            configuration.TypesToIncludeInScan(GetTypesScopedByTestClass(endpointConfiguration));
+            configuration.TypesToIncludeInScan(endpointConfiguration.GetTypesScopedByTestClass());
             configuration.EnableInstallers();
 
             configuration.DisableFeature<TimeoutManager>();
@@ -44,41 +40,6 @@
             return Task.FromResult(configuration);
         }
 
-        static IEnumerable<Type> GetTypesScopedByTestClass(EndpointCustomizationConfiguration endpointConfiguration)
-        {
-            var assemblies = new AssemblyScanner().GetScannableAssemblies();
-
-            var assembliesToScan = assemblies.Assemblies
-                //exclude acceptance tests by default
-                .Where(a => a != Assembly.GetExecutingAssembly()).ToList();
-            var types = assembliesToScan
-                .SelectMany(a => a.GetTypes());
-
-            types = types.Union(GetNestedTypeRecursive(endpointConfiguration.BuilderType.DeclaringType, endpointConfiguration.BuilderType));
-
-            types = types.Union(endpointConfiguration.TypesToInclude);
-
-            return types.Where(t => !endpointConfiguration.TypesToExclude.Contains(t)).ToList();
-        }
-
-        static IEnumerable<Type> GetNestedTypeRecursive(Type rootType, Type builderType)
-        {
-            if (rootType == null)
-            {
-                throw new InvalidOperationException("Make sure you nest the endpoint infrastructure inside the TestFixture as nested classes");
-            }
-
-            yield return rootType;
-
-            if (typeof(IEndpointConfigurationFactory).IsAssignableFrom(rootType) && rootType != builderType)
-            {
-                yield break;
-            }
-
-            foreach (var nestedType in rootType.GetNestedTypes(BindingFlags.Instance | BindingFlags.Static | BindingFlags.Public | BindingFlags.NonPublic).SelectMany(t => GetNestedTypeRecursive(t, builderType)))
-            {
-                yield return nestedType;
-            }
-        }
+        
     }
 }
